Create shared test fixture module
